added support for realtime notification with Pusher, decoupled the Scheduler
I added a Pusher Notificator to the javascript assets, that might
accept subscription to several Backbone collections. In each Model View,
there is a subsequent bound to its related events, who trigger the
render() callback. In each Model View, there is a subsequent bound to
its related events, who trigger the render() callback.
Regarding the Scheduler, it is now decoupled from the Chick Document and
put in an Observer.